re-added logic to ownership change notification component
