+++ src/app/features/notifications/notification/notification.component.ts
@@ -59,11 +59,6 @@ export class NotificationComponent implements OnInit, OnDestroy {
           ? details.typeContent.subEstablishmentName
           : details.typeContent.parentEstablishmentName;
 
-      if (details.typeContent.approvalStatus === 'APPROVED') {
-        this.isWorkPlaceIsRequester = this.workplace.name !== this.ownerShipRequestedFrom;
-      } else {
-        this.isWorkPlaceIsRequester = this.workplace.name === this.ownerShipRequestedFrom;
-      }
       this.displayActionButtons =
         details.typeContent.approvalStatus === 'REQUESTED' || details.typeContent.approvalStatus === 'CANCELLED';
     });
